requirements.txt
# Autonomous Threat Hunter — Python Dependencies
# Install: pip install -r requirements.txt

# ── Phase 1-2: Setup, Testing, AWS Collection ──
requests>=2.31.0
boto3>=1.34.0
python-dotenv>=1.0.0

# ── Phase 2: Elasticsearch client ──
elasticsearch>=8.13.0,<9.0.0

# ── Phase 2: Scheduler ──
celery>=5.3.0
redis>=5.0.0

# ── Phase 4-5: Correlation + MITRE ──
networkx>=3.2.0

# ── Phase 6: AI Agent + RAG ──
langchain>=0.2.0
langgraph>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0

# ── Phase 9: Dashboard ──
streamlit>=1.35.0

# ── Phase 8: FastAPI backend ──
fastapi>=0.111.0
uvicorn>=0.30.0
